:mouse: 修正した

+++ Utility.ps1
@@ -5,40 +5,10 @@ function Wait-PressAnyKeyToContinue {
     $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
 }
 
-function Rename-ItemWithModifiedDate {
-    param([parameter(Mandatory)][object[]]$ItemList)
-<#
-    .SYNOPSIS
-    ファイル名先頭8文字に日付がなければファイル名の先頭に日付を加える。
-#>
-
-    foreach ($item in $ItemList) {
-        if ($item -is [String]) {
-            $item = Get-Item $item
-        }
-
-        # ファイル名に日付が含まれるか確認する
-        if ($item.Name -match "^(?<Year>\d{4})(?<Month>\d{2})(?<Day>\d{2})") {
-            continue
-        }
-
-        # ファイルの先頭の日付を得る
-        $prefix = Get-Date -Date $item.LastWriteTime -Format "yyyyMMdd_"
-
-        # リネームする
-        $newFilename = $prefix + $item.Name
-        $newFullname = Join-Path $item.Directory.FullName $newFilename
-        Write-Information "---"
-        Write-Information "src: $newFullname"
-        Write-Information "dest: $($item.FullName)"
-        Rename-Item -NewName $newFullname -LiteralPath $item
-    }
-}
-
 function Get-ArchiveParentDir {
     [OutputType([String])]
     Param([String]$ParentFullName)
-<#
+    <#
     .SYNOPSIS
     指定ファイルまたはディレクトリに対するアーカイブディレクトリを得る
 
@@ -50,10 +20,10 @@ function Get-ArchiveParentDir {
     [String]$parent = $ParentFullName
     Do {
         $leaf = (Split-Path $parent -Leaf)
+        $parent = Split-Path $parent -Parent
         If ($leaf -eq $ARCHIVE_DIR_NAME) {
             return $parent
         }
-        $parent = Split-Path $parent -Parent
     } While ($parent -ne "")
 
     $ParentFullName
@@ -62,7 +32,7 @@ function Get-ArchiveParentDir {
 function Test-DirHasArchive {
     [OutputType([String])]
     Param([String]$FullName)
-<#
+    <#
     .SYNOPSIS
     ARCHIVE_DIR_NAME が含まれるか確認する
 #>
@@ -82,9 +52,9 @@ function Test-DirHasArchive {
 function Get-ArchivePath {
     [OutputType([String])]
     Param(
-        [System.DateTime][parameter(mandatory=$true)]$ItemDate,
+        [System.DateTime][parameter(mandatory = $true)]$ItemDate,
         [System.DateTime]$Date = (Get-Date))
-<#
+    <#
     .SYNOPSIS
     ARCHIVE_DIR_NAME の親ディレクトリから、ファイルが格納されるべきディレクトリまでのパスを返す。
 
@@ -98,8 +68,7 @@ function Get-ArchivePath {
     If ($ItemDate.Year -lt $Date.Year) {
         $parts += @(Get-Date $ItemDate -Format "yyyy")
         $parts += @(Get-Date $ItemDate -Format "yyyyMM")
-    }
-    ElseIf ($ItemDate.Year -eq $Date.Year -and $ItemDate.Month -lt $Date.Month) {
+    } ElseIf ($ItemDate.Year -eq $Date.Year -and $ItemDate.Month -lt $Date.Month) {
         $parts += @(Get-Date $ItemDate -Format "yyyyMM")
     }
 
@@ -111,9 +80,75 @@ function Get-ArchivePath {
     $ret
 }
 
+function Get-ItemNamePrefixDate {
+    <#
+    .SYNOPSIS
+    $Item.Name の先頭8文字による日付を得る
+#>
+    param([parameter(Mandatory)]$Item)
+    If ($Item.Name -match "^(?<Year>\d{4})(?<Month>\d{2})(?<Day>\d{2})\D") {
+        try {
+            Get-Date -Year $Matches.Year -Month $Matches.Month -Day $Matches.Day -Hour 0 -Minute 0 -Second 0 -Millisecond 0
+        } catch {
+            $null
+        }
+    } else {
+        $null
+    }
+}
+
+function Test-ItemNamePrefixDate {
+    <#
+    .SYNOPSIS
+    $Item.Name の先頭8文字が日付であるか判定する
+#>
+    param([parameter(Mandatory)]$Item)
+    $null -ne (Get-ItemNamePrefixDate $Item)
+}
+
+function Get-ItemArchiveParentDirectory {
+    param($Item)
+    If ($Item -is [System.IO.FileInfo]) {
+        $parentFullName = $Item.Directory.FullName
+    } ElseIf ($Item -is [System.IO.DirectoryInfo]) {
+        $parentFullName = $Item.Parent.FullName
+    }
+    Get-ArchiveParentDir -ParentFullName $parentFullName
+}
+
+function Rename-ItemWithModifiedDate {
+    param([parameter(Mandatory)][object[]]$ItemList)
+    <#
+    .SYNOPSIS
+    ファイル名先頭8文字に日付がなければファイル名の先頭に日付を加える。
+#>
+
+    foreach ($item in $ItemList) {
+        if ($item -is [String]) {
+            $item = Get-Item $item
+        }
+
+        # ファイル名に日付が含まれるなら処理しない
+        if (Test-ItemNamePrefixDate $item) {
+            continue
+        }
+
+        # 変更内容を出力する
+        [String]$prefix = Get-Date -Date $item.LastWriteTime -Format "yyyyMMdd_"
+        $newFilename = $prefix + $item.Name
+        $newFullname = Join-Path $item.Directory.FullName $newFilename
+        Write-Information "---"
+        Write-Information "src: $($item.FullName)"
+        Write-Information "dst: $newFullname"
+
+        # リネームする
+        Rename-Item -NewName $newFullname -LiteralPath $item
+    }
+}
+
 function Move-ItemToArchiveDirByDate {
     param([parameter(Mandatory)][object[]]$ItemList)
-<#
+    <#
     .SYNOPSIS
     ARCHIVE_DIR_NAME の親ディレクトリから、ファイルが格納されるべきディレクトリまでのパスを返す。
 
@@ -128,40 +163,29 @@ function Move-ItemToArchiveDirByDate {
             $Item = Get-Item $Item
         }
 
-        # ファイル名に日付が含まれるか確認する
-        If (-not ($Item.Name -match "^(?<Year>\d{4})(?<Month>\d{2})(?<Day>\d{2})")) {
+        # 移動先のディレクトリがなければ作る
+        $archiveParentDir = Get-ItemArchiveParentDirectory $Item
+        $date = Get-ItemNamePrefixDate $Item
+        if ($null -eq $date) {
             throw "ファイル名の先頭8文字が整数ではなかった: $($Item.FullName)"
         }
-        [System.DateTime]$date = Get-Date -Year $Matches.Year -Month $Matches.Month -Day $Matches.Day
-
-        # アーカイブフォルダの親ディレクトリを得る
-        If ($Item -is [System.IO.FileInfo]) {
-            $parentFullName = $Item.Directory.FullName
-        }
-        ElseIf ($Item -is [System.IO.DirectoryInfo]) {
-            $parentFullName = $Item.Parent.FullName
-        }
-        $archiveParentDir = Get-ArchiveParentDir -ParentFullName $parentFullName
-
-        # 移動先のディレクトリがなければ作る
-        [String]$archivePath = Get-ArchivePath -ItemDate $date
-        [String]$destDir = Join-Path $archiveParentDir $archivePath
+        $archivePath = Get-ArchivePath -ItemDate $date
+        $destDir = Join-Path $archiveParentDir $archivePath
         if (-not (Test-Path $destDir -PathType Container)) {
             New-Item $destDir -ItemType Container | Out-Null
         }
 
-        # 移動する
+        # 変更内容を出力する
         Write-Information "---"
         Write-Information "src: $($Item.FullName)"
-        Write-Information "dest: $($destDir)"
+        Write-Information "dst: $($destDir)"
+
+        # 移動する
         Move-Item $Item -Destination $destDir
 
         # 移動元のフォルダが $ARCHIVE_DIR_NAME 配下で、アイテム数がゼロならディレクトリを消す
-        if (Test-DirHasArchive $parentFullName) {
-            $m = Get-ChildItem $parentFullName | Measure-Object
-            if ($m.Count -eq 0) {
-                Remove-Item -LiteralPath $parentFullName
-            }
+        if ((Test-DirHasArchive $parentFullName) -and ((Get-ChildItem $parentFullName | Measure-Object).Count -eq 0)) {
+            Remove-Item -LiteralPath $parentFullName
         }
     }
 }
